+++ @navikt/core/react/src/table/index.ts
@@ -1,4 +1,3 @@
-export { default as Table, type TableProps, type SortState } from "./Table";
 export { type BodyProps } from "./Body";
 export { type ColumnHeaderProps } from "./ColumnHeader";
 export { type DataCellProps } from "./DataCell";
@@ -6,3 +5,5 @@ export { type ExpandableRowProps } from "./ExpandableRow";
 export { type HeaderProps } from "./Header";
 export { type HeaderCellProps } from "./HeaderCell";
 export { type RowProps } from "./Row";
+export { default as Table, type TableProps } from "./Table";
+export { type SortState } from "./types";
