+++ @navikt/core/react/src/table/types.ts
@@ -1,0 +1,4 @@
+export interface SortState {
+  orderBy: string;
+  direction: "ascending" | "descending";
+}
